+++ akshare/__init__.py
@@ -2040,9 +2040,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.6.19 fix: fix stock_comment_detail_zlkp_jgcyd_em interface
 1.6.20 fix: fix stock_board_industry_name_em and code_id_map_em interface
 1.6.21 fix: fix baidu_search_index interface
+1.6.22 add: add stock_kc_a_spot_em interface
 """
 
-__version__ = "1.6.21"
+__version__ = "1.6.22"
 __author__ = "AKFamily"
 
 import sys
@@ -2605,6 +2606,7 @@ from akshare.stock_feature.stock_hist_em import (
     stock_zh_a_spot_em,
     stock_bj_a_spot_em,
     stock_new_a_spot_em,
+    stock_kc_a_spot_em,
     stock_sh_a_spot_em,
     stock_sz_a_spot_em,
     stock_zh_b_spot_em,