+++ docs/changelog.md
@@ -29,6 +29,10 @@
 
 ## 更新说明
 
+1.6.22 add: add stock_kc_a_spot_em interface
+
+    1. 新增 stock_kc_a_spot_em 接口，获取科创板行情
+
 1.6.21 fix: fix baidu_search_index interface
 
     1. 修复 baidu_search_index 接口
@@ -771,6 +775,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.22 add: add stock_kc_a_spot_em interface
 
 1.6.21 fix: fix baidu_search_index interface